Remove old cody web
