+++ client/shared/src/settings/settings.tsx
@@ -311,7 +311,6 @@ const defaultFeatures: SettingsExperimentalFeatures = {
     isInitialized: true,
     searchQueryInput: 'v2',
     newSearchResultFiltersPanel: true,
-    newCodyWeb: true,
 }
 
 /**
