+++ client/web/BUILD.bazel
@@ -204,13 +204,9 @@ ts_project(
         "src/cody/chat/CodyChatPage.tsx",
         "src/cody/chat/CodyPageIcon.tsx",
         "src/cody/chat/index.tsx",
-        "src/cody/chat/new-chat/NewCodyChatPage.tsx",
-        "src/cody/chat/new-chat/components/chat-ui/ChatUi.tsx",
-        "src/cody/chat/old-chat/CodyChatPage.tsx",
         "src/cody/codyProRoutes.tsx",
         "src/cody/codyRoutes.tsx",
-        "src/cody/components/ChatUI/ChatUi.tsx",
-        "src/cody/components/ChatUI/index.tsx",
+        "src/cody/components/ChatUi.tsx",
         "src/cody/components/CodeMirrorEditor.ts",
         "src/cody/components/CodyAlert.tsx",
         "src/cody/components/CodyContainer.tsx",
@@ -218,17 +214,8 @@ ts_project(
         "src/cody/components/CodyLogo.tsx",
         "src/cody/components/CodyProBadgeDeck.tsx",
         "src/cody/components/FileContentEditor.ts",
-        "src/cody/components/GettingStarted.tsx",
-        "src/cody/components/HistoryList/HistoryList.tsx",
-        "src/cody/components/HistoryList/index.tsx",
         "src/cody/components/PageHeaderIcon.tsx",
         "src/cody/components/RepoContainerEditor.ts",
-        "src/cody/components/ScopeSelector/Callout.tsx",
-        "src/cody/components/ScopeSelector/RepositoriesSelectorPopover.tsx",
-        "src/cody/components/ScopeSelector/ScopeSelector.tsx",
-        "src/cody/components/ScopeSelector/backend.ts",
-        "src/cody/components/ScopeSelector/index.tsx",
-        "src/cody/components/ScopeSelector/useRepoSuggestions.ts",
         "src/cody/dashboard/CodyDashboardPage.tsx",
         "src/cody/invites/AcceptInviteBanner.tsx",
         "src/cody/invites/InviteUsers.tsx",
@@ -264,11 +251,9 @@ ts_project(
         "src/cody/management/subscription/new/CodyProCheckoutForm.tsx",
         "src/cody/management/subscription/new/NewCodyProSubscriptionPage.tsx",
         "src/cody/sidebar/CodySidebar.tsx",
+        "src/cody/sidebar/CodySidebarWebChat.tsx",
         "src/cody/sidebar/Provider.tsx",
         "src/cody/sidebar/index.tsx",
-        "src/cody/sidebar/new-cody-sidebar/NewCodySidebar.tsx",
-        "src/cody/sidebar/new-cody-sidebar/NewCodySidebarWebChat.tsx",
-        "src/cody/sidebar/old-cody-sidebar/CodySidebar.tsx",
         "src/cody/sidebar/useSidebarSize.tsx",
         "src/cody/subscription/CodySubscriptionPage.tsx",
         "src/cody/subscription/queries.tsx",
@@ -280,10 +265,6 @@ ts_project(
         "src/cody/useCodyChat.tsx",
         "src/cody/useCodyIgnore.tsx",
         "src/cody/util.ts",
-        "src/cody/widgets/CodyRecipesWidget.tsx",
-        "src/cody/widgets/components/Recipe.tsx",
-        "src/cody/widgets/components/RecipeAction.tsx",
-        "src/cody/widgets/components/Recipes.tsx",
         "src/communitySearchContexts/Backstage.tsx",
         "src/communitySearchContexts/CommunitySearchContextPage.tsx",
         "src/communitySearchContexts/HomepageConfig.ts",
@@ -1325,7 +1306,6 @@ ts_project(
         "src/repo/blob/codemirror/search.ts",
         "src/repo/blob/codemirror/search/SearchPanel.tsx",
         "src/repo/blob/codemirror/static-highlights.tsx",
-        "src/repo/blob/codemirror/tooltips/CodyTooltip.tsx",
         "src/repo/blob/codemirror/tooltips/HovercardView.tsx",
         "src/repo/blob/codemirror/tooltips/LoadingTooltip.ts",
         "src/repo/blob/codemirror/tooltips/TemporaryTooltip.ts",
