--- client/web/src/cody/components/ChatUI/index.tsx
@@ -1,1 +1,0 @@
-export * from './ChatUi'
