--- client/web/src/cody/components/GettingStarted.module.scss
@@ -1,93 +1,0 @@
-.container {
-    flex: 1;
-    background: var(--color-bg-1);
-    overflow: auto;
-}
-
-.icons {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-}
-
-.icon-section {
-    margin-bottom: 3rem !important;
-    display: flex;
-    align-items: flex-end;
-    justify-content: center;
-}
-
-.section {
-    background: var(--body-bg);
-    padding: 1rem;
-    border-radius: 4px;
-}
-
-.greeting {
-    &-container {
-        position: relative;
-    }
-
-    &-text {
-        color: var(--white);
-        position: absolute;
-        top: 1rem;
-        left: 2rem;
-    }
-}
-
-.cody-icon-container {
-    grid-column: 1 / 1;
-    grid-row: 2 / 2;
-    justify-self: end;
-}
-
-.cody-icon {
-    // stylelint-disable-next-line declaration-property-unit-allowed-list
-    height: 32px !important;
-    // stylelint-disable-next-line declaration-property-unit-allowed-list
-    width: 32px !important;
-}
-
-.radio-wrapper {
-    margin-bottom: 0.75rem;
-}
-
-.scope-selector {
-    &-wrapper {
-        padding-left: var(--form-check-input-gutter);
-    }
-
-    &-warning,
-    &-warning-link {
-        margin-top: 0.5rem;
-        margin-bottom: 0;
-        color: var(--warning-3);
-    }
-
-    &-warning-link {
-        text-decoration: underline;
-    }
-
-    &-warning-link:hover {
-        color: var(--warning);
-    }
-}
-
-.divider {
-    margin: 0.75rem 0;
-}
-
-.example-wrapper {
-    margin-top: 0.625rem;
-}
-
-.hint {
-    &-title {
-        font-weight: 500;
-        margin-bottom: 0.325rem;
-    }
-
-    &-text {
-        margin-bottom: 0.625rem;
-    }
-}
