--- client/web/src/cody/components/HistoryList/HistoryList.module.scss
@@ -1,27 +1,0 @@
-.history-item {
-    display: flex;
-    flex-direction: column;
-    margin-bottom: 0;
-    padding: var(--spacing);
-    border: none;
-    cursor: pointer;
-    border-radius: 0.25rem;
-    background: var(--body-bg);
-
-    &:hover {
-        background: var(--color-bg-1);
-    }
-
-    &.selected,
-    &.selected:hover {
-        background: var(--color-bg-2);
-    }
-
-    .delete-button {
-        visibility: hidden;
-    }
-
-    &:hover .delete-button {
-        visibility: visible;
-    }
-}
