--- client/web/src/cody/components/HistoryList/index.tsx
@@ -1,1 +1,0 @@
-export { HistoryList } from './HistoryList'
