--- client/web/src/cody/components/ScopeSelector/Callout.module.scss
@@ -1,64 +1,0 @@
-@keyframes slideFadeIn {
-    0% {
-        opacity: 0;
-        transform: translateX(0.625rem);
-    }
-    100% {
-        opacity: 1;
-        transform: translateX(0);
-    }
-}
-
-.wrapper {
-    width: auto;
-    max-width: 26rem;
-    overflow: hidden;
-    margin: 0;
-    animation: slideFadeIn 0.5s ease-out forwards;
-    position: absolute;
-    bottom: 2rem;
-    left: 1rem;
-}
-
-.box {
-    background-color: var(--brand-secondary);
-    color: var(--white);
-    border-radius: 0.375rem;
-    box-shadow: 0 0.125rem 0.375rem rgba(0, 0, 0, 0.3);
-    padding: 0.8125rem 1rem;
-}
-
-.header-elements {
-    display: flex;
-    align-items: top;
-    gap: 0.28rem;
-}
-
-.header {
-    font-size: 1.15rem;
-    font-weight: 400;
-    display: flex;
-    justify-content: space-between;
-    align-self: center;
-}
-
-.header .close-button {
-    position: relative;
-    margin-right: -0.5rem;
-    margin-top: -0.5rem;
-}
-
-.content {
-    word-break: break-all;
-    white-space: normal;
-}
-
-.tail {
-    position: relative;
-    width: 0;
-    height: 0;
-    border-left: 9px solid transparent;
-    border-right: 9px solid transparent;
-    border-top: 9px solid var(--brand-secondary);
-    left: 1.875rem;
-}
