--- client/web/src/cody/components/ScopeSelector/index.tsx
@@ -1,1 +1,0 @@
-export { ScopeSelector } from './ScopeSelector'
