+++ client/web/src/cody/sidebar/CodySidebar.module.scss
