--- client/web/src/cody/widgets/components/Recipes.module.scss
@@ -1,30 +1,0 @@
-.recipes-wrapper {
-    display: flex;
-    overflow-x: auto;
-    background-color: var(--body-bg);
-    border-radius: 6px;
-    padding: 0.18rem 0.06rem;
-    align-items: center;
-    overflow: hidden;
-    filter: drop-shadow(1px 2px 3px rgba(36, 41, 54, 0.2));
-    user-select: none;
-}
-
-.recipe-wrapper {
-    background-color: var(--input-bg);
-    border-radius: 4px;
-    border: 1px solid var(--border-color-2);
-    padding: 0.06rem 0.5rem;
-    /* stylelint-disable-next-line declaration-property-unit-allowed-list */
-    margin: 0 3px;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    font-size: 0.8rem;
-    display: inline-block;
-    flex-shrink: 0;
-    font-weight: 400;
-}
-
-.recipe-menu-wrapper {
-    font-size: 0.83rem;
-}
