--- client/web/src/cody/widgets/components/Recipes.tsx
@@ -1,14 +1,0 @@
-import React from 'react'
-
-import classNames from 'classnames'
-
-import { AskCodyIcon } from '@sourcegraph/cody-ui'
-
-import styles from './Recipes.module.scss'
-
-export const Recipes: React.FC<React.PropsWithChildren<{}>> = ({ children }) => (
-    <div className={classNames(styles.recipesWrapper)}>
-        <AskCodyIcon />
-        {children}
-    </div>
-)
