+++ schema/settings.schema.json
@@ -29,14 +29,6 @@
             "pointer": true
           }
         },
-        "newCodyWeb": {
-          "description": "Enables new experimental Cody Web UI",
-          "type": "boolean",
-          "default": false,
-          "!go": {
-            "pointer": true
-          }
-        },
         "newSearchResultsUI": {
           "description": "Enables new experimental search results UI, such as preview panel feature and updated search and filter layouts.",
           "type": "boolean",
